✨ feat: rename Islands to Rounded tabs and add rounded provider replacing old one
