+++ src/main/java/krasa/editorGroups/settings/EditorGroupsSettings.kt
@@ -85,7 +85,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     var tabsPlacement: Int by property(SwingConstants.TOP)
 
     // Islands
-    var isIslands: Boolean by property(true)
+    var isRoundedTabs: Boolean by property(true)
   }
 
   @EditorGroupSetting([EditorGroupSetting.Category.REGEX, EditorGroupSetting.Category.GROUPS])
@@ -257,10 +257,10 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     }
 
   @EditorGroupSetting([EditorGroupSetting.Category.UI, EditorGroupSetting.Category.TABS])
-  var isIslands: Boolean
-    get() = state.isIslands
+  var isRoundedTabs: Boolean
+    get() = state.isRoundedTabs
     set(value) {
-      state.isIslands = value
+      state.isRoundedTabs = value
     }
 
   fun fireChanged() {
@@ -295,7 +295,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     clone.reuseCurrentTab = this.reuseCurrentTab
     clone.tabSizeLimit = this.tabSizeLimit
     clone.tabsPlacement = this.tabsPlacement
-    clone.isIslands = this.isIslands
+    clone.isRoundedTabs = this.isRoundedTabs
     return clone
   }
 
@@ -324,7 +324,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     this.reuseCurrentTab = state.reuseCurrentTab
     this.tabSizeLimit = state.tabSizeLimit
     this.tabsPlacement = state.tabsPlacement
-    this.isIslands = state.isIslands
+    this.isRoundedTabs = state.isRoundedTabs
     this.fireChanged()
   }
 
@@ -363,7 +363,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     this.tabsPlacement = SwingConstants.TOP
     this.isCustomFont = false
     this.customFont = DEFAULT_FONT
-    this.isIslands = true
+    this.isRoundedTabs = true
     this.fireChanged()
   }
 
@@ -398,7 +398,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     if (reuseCurrentTab != other.reuseCurrentTab) return false
     if (tabSizeLimit != other.tabSizeLimit) return false
     if (tabsPlacement != other.tabsPlacement) return false
-    if (isIslands != other.isIslands) return false
+    if (this@EditorGroupsSettings.isRoundedTabs != other.isRoundedTabs) return false
 
     return true
   }
@@ -428,7 +428,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     result = 31 * result + reuseCurrentTab.hashCode()
     result = 31 * result + tabSizeLimit
     result = 31 * result + tabsPlacement
-    result = 31 * result + isIslands.hashCode()
+    result = 31 * result + this@EditorGroupsSettings.isRoundedTabs.hashCode()
     return result
   }
 
@@ -458,7 +458,7 @@ class EditorGroupsSettings : SimplePersistentStateComponent<EditorGroupsSettings
     |reuseCurrentTab=$reuseCurrentTab,
     |tabSizeLimit=$tabSizeLimit,
     |tabsPlacement=$tabsPlacement,
-    |isIslands=$isIslands,
+    |isRoundedTabs=${this@EditorGroupsSettings.isRoundedTabs},
     )
   """.trimMargin()
