--- src/main/java/krasa/editorGroups/tabs2/impl/islands/IslandsTabsPainterProvider.kt
@@ -1,7 +1,0 @@
-package krasa.editorGroups.tabs2.impl.islands
-
-import krasa.editorGroups.tabs2.impl.painter.EditorGroupsTabPainterAdapter
-
-internal class IslandsTabsPainterProvider : KrTabsPainterProvider() {
-  override fun createTabPainter(): EditorGroupsTabPainterAdapter = IslandsTabPainterAdapter()
-}
