+++ src/main/java/krasa/editorGroups/tabs2/impl/rounded/KrRoundedTabsPainterProvider.kt
@@ -1,13 +1,13 @@
-package krasa.editorGroups.tabs2.impl.islands
+package krasa.editorGroups.tabs2.impl.rounded
 
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.components.serviceOrNull
 import krasa.editorGroups.tabs2.impl.painter.EditorGroupsTabPainterAdapter
 
-open class KrTabsPainterProvider {
+open class KrRoundedTabsPainterProvider {
   companion object {
     @JvmStatic
-    fun getInstance(): KrTabsPainterProvider? = ApplicationManager.getApplication()?.serviceOrNull()
+    fun getInstance(): KrRoundedTabsPainterProvider? = ApplicationManager.getApplication()?.serviceOrNull()
   }
 
   open fun createTabPainter(): EditorGroupsTabPainterAdapter? = null
