+++ src/main/resources/META-INF/plugin.xml
@@ -119,8 +119,9 @@
         <!-- Theme -->
         <themeMetadataProvider path="/themes/metadata/EditorGroups.themeMetadata.json"/>
 
-        <applicationService serviceInterface="krasa.editorGroups.tabs2.impl.islands.KrTabsPainterProvider"
-                            serviceImplementation="krasa.editorGroups.tabs2.impl.islands.IslandsTabsPainterProvider"/>
+        <!-- Rounded Tabs (Islands) -->
+        <applicationService serviceInterface="krasa.editorGroups.tabs2.impl.rounded.KrRoundedTabsPainterProvider"
+                            serviceImplementation="krasa.editorGroups.tabs2.impl.rounded.RoundedTabsPainterProvider"/>
     </extensions>
 
     <!--suppress PluginXmlValidity -->
